Multiwinner: change ints referring to candidates from int to size_t.
Also add some patches to range_stv and qrange_stv so they remove ballots
with negative weight. I'm not sure if this is the right thing to do, but
at least I can test them without exceptions being thrown now. Fix later.

+++ src/multiwinner/auction.h
@@ -15,12 +15,12 @@ class r_auction : public multiwinner_method {
 			std::vector<double> & account_balance,
 			const election_t & ballots,
 			double minimum, double maximum,
-			int num_candidates, bool cumulative) const;
+			size_t num_candidates, bool cumulative) const;
 
 		bool cumul_setting;
 
 	public:
-		std::list<int> get_council(int council_size, int num_candidates,
+		std::list<size_t> get_council(size_t council_size, size_t num_candidates,
 			const election_t & ballots) const;
 
 		r_auction(bool cumul_in) {
