+++ src/multiwinner/shuntsstv.h
@@ -16,8 +16,8 @@
 class SchulzeSTV : public multiwinner_method {
 
 	public:
-		std::list<int> get_council(int council_size,
-			int num_candidates, const election_t & ballots) const;
+		std::list<size_t> get_council(size_t council_size,
+			size_t num_candidates, const election_t & ballots) const;
 
 		std::string name() const {
 			return ("Schulze STV");
